fix(wizard): seed yoloMode from profile config on mount
The web session wizard hardcoded `yoloMode: false` and only ran the
profile-defaults loader when the user actively swapped the profile
picker. The picker is hidden when there is one or fewer profiles
(`AgentStep.tsx`), so single-profile users with `yolo_mode_default =
true` never had their preference applied; sessions launched in default
permission mode despite the TUI honoring the same setting.

The mount effect now resolves the effective profile (prefill > the
server's active profile from `/api/profiles` > unresolved global
config) and dispatches `APPLY_PROFILE_DEFAULTS` with the same merge
rules `AgentStep.handleProfileChange` already uses, so the wizard
opens with the profile-resolved yolo, sandbox, tool, and extra-env
values. Explicit prefill values still win, and a `skipIfDirty` flag
on the action keeps a slow `/api/settings` response from clobbering
edits the user made between mount and the fetch resolving.

The reducer moves to its own file so it can be unit-tested without
mounting React (the wizard imports `fetchSettings`/`fetchProfiles`,
which would force a network mock for every test) and so the wizard
file stays a pure component module under
`react-refresh/only-export-components`.

Fixes #1142
