+++ web/src/components/session-wizard/SessionWizard.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo, useReducer } from "react";
-import type { AgentInfo, GroupInfo, ProfileInfo, CreateSessionRequest, SessionResponse } from "../../lib/types";
+import type { CreateSessionRequest, SessionResponse } from "../../lib/types";
 import { fetchAgents, fetchGroups, fetchDockerStatus, fetchProfiles, fetchSettings, createSession } from "../../lib/api";
 import { ACP_CAPABLE_TOOLS } from "../../lib/acpCapableTools";
 import { toastBus } from "../../lib/toastBus";
@@ -9,127 +9,10 @@ import { ProjectStep } from "./steps/ProjectStep";
 import { SessionStep } from "./steps/SessionStep";
 import { AgentStep } from "./steps/AgentStep";
 import { ReviewStep } from "./steps/ReviewStep";
-import { applyBranchOverride, getSubmittedBranch, slugifyBranch } from "./sessionNames";
+import { getSubmittedBranch } from "./sessionNames";
+import { initialData, reducer, type WizardData } from "./wizardReducer";
 
-export interface WizardData {
-  path: string;
-  title: string;
-  worktreeBranch: string;
-  worktreeBranchDirty: boolean;
-  useWorktree: boolean;
-  /** When true, attach to an existing branch's worktree (`create_new_branch: false`
-   *  on the API). Mirrors the TUI new-session "Attach to existing branch"
-   *  toggle (`src/tui/dialogs/new_session/render.rs:851`). See #969. */
-  attachExisting: boolean;
-  /** Optional base branch for the new worktree branch. Empty string =
-   *  use the project's default branch. Lives under "Advanced" in the
-   *  session step. See #948. */
-  baseBranch: string;
-  group: string;
-  tool: string;
-  profile: string;
-  yoloMode: boolean;
-  sandboxEnabled: boolean;
-  sandboxImage: string;
-  extraEnv: string[];
-  /** Additional repo paths to include in the multi-repo workspace.
-   *  Free-text paths and registered project paths flow into the same list. */
-  extraRepoPaths: string[];
-  advancedEnabled: boolean;
-  customInstruction: string;
-  extraArgs: string;
-  commandOverride: string;
-  /** Tracks whether the user has manually edited fields after a profile selection */
-  profileDirty: boolean;
-  [key: string]: unknown;
-}
-
-interface WizardState {
-  currentStep: number;
-  data: WizardData;
-  isSubmitting: boolean;
-  error: string | null;
-  agents: AgentInfo[];
-  groups: GroupInfo[];
-  profiles: ProfileInfo[];
-  dockerAvailable: boolean;
-}
-
-type Action =
-  | { type: "SET_FIELD"; field: string; value: unknown }
-  | { type: "SET_STEP"; step: number }
-  | { type: "SUBMIT_START" }
-  | { type: "SUBMIT_ERROR"; error: string }
-  | { type: "SUBMIT_SUCCESS" }
-  | { type: "SET_AGENTS"; agents: AgentInfo[] }
-  | { type: "SET_GROUPS"; groups: GroupInfo[] }
-  | { type: "SET_PROFILES"; profiles: ProfileInfo[] }
-  | { type: "SET_DOCKER"; available: boolean }
-  | { type: "APPLY_PROFILE_DEFAULTS"; yoloMode: boolean; sandboxEnabled: boolean; tool: string; extraEnv: string[] };
-
-const initialData: WizardData = {
-  path: "", title: "", worktreeBranch: "", worktreeBranchDirty: false,
-  useWorktree: true, attachExisting: false, baseBranch: "",
-  group: "", tool: "claude", profile: "",
-  yoloMode: false, sandboxEnabled: false, sandboxImage: "", extraEnv: [],
-  extraRepoPaths: [],
-  advancedEnabled: false, profileDirty: false,
-  customInstruction: "", extraArgs: "", commandOverride: "",
-};
-
-function reducer(state: WizardState, action: Action): WizardState {
-  switch (action.type) {
-    case "SET_FIELD": {
-      const newData = { ...state.data, [action.field]: action.value };
-      if (action.field === "title" && !state.data.worktreeBranchDirty) {
-        newData.worktreeBranch = slugifyBranch(String(action.value));
-      }
-      if (action.field === "worktreeBranch") {
-        const override = applyBranchOverride(
-          String(newData.title),
-          String(action.value),
-        );
-        newData.worktreeBranch = override.worktreeBranch;
-        newData.worktreeBranchDirty = override.worktreeBranchDirty;
-      }
-      // Mark as dirty when user manually edits agent-step fields after a profile was chosen
-      if (state.data.profile && ["yoloMode", "sandboxEnabled", "tool", "extraEnv"].includes(action.field)) {
-        newData.profileDirty = true;
-      }
-      return { ...state, data: newData, error: null };
-    }
-    case "SET_STEP":
-      return { ...state, currentStep: action.step };
-    case "SUBMIT_START":
-      return { ...state, isSubmitting: true, error: null };
-    case "SUBMIT_ERROR":
-      return { ...state, isSubmitting: false, error: action.error };
-    case "SUBMIT_SUCCESS":
-      return { ...state, isSubmitting: false };
-    case "SET_AGENTS":
-      return { ...state, agents: action.agents };
-    case "SET_GROUPS":
-      return { ...state, groups: action.groups };
-    case "SET_PROFILES":
-      return { ...state, profiles: action.profiles };
-    case "SET_DOCKER":
-      return { ...state, dockerAvailable: action.available };
-    case "APPLY_PROFILE_DEFAULTS":
-      return {
-        ...state,
-        data: {
-          ...state.data,
-          yoloMode: action.yoloMode,
-          sandboxEnabled: action.sandboxEnabled,
-          tool: action.tool || state.data.tool,
-          extraEnv: action.extraEnv,
-          profileDirty: false,
-        },
-      };
-    default:
-      return state;
-  }
-}
+export type { WizardData } from "./wizardReducer";
 
 // Wizard: project path → session (title + worktree) → agent → review
 function computeSteps(_data: WizardData): StepDef[] {
@@ -193,15 +76,55 @@ export function SessionWizard({ onClose, onCreated, prefill, cockpitMasterEnable
   useEffect(() => {
     fetchAgents().then((a) => dispatch({ type: "SET_AGENTS", agents: a }));
     fetchGroups().then((g) => dispatch({ type: "SET_GROUPS", groups: g }));
-    fetchProfiles().then((p) => dispatch({ type: "SET_PROFILES", profiles: p }));
     fetchDockerStatus().then((d) => dispatch({ type: "SET_DOCKER", available: d.available }));
-    fetchSettings().then((s) => {
-      if (s) {
+
+    // Seed the wizard with the resolved (global + active profile) defaults so
+    // single-profile users get yolo_mode_default and friends without ever
+    // touching the profile picker. The picker is hidden when
+    // profiles.length <= 1 (`AgentStep.tsx`), so its onChange-driven
+    // `APPLY_PROFILE_DEFAULTS` path never fires and the wizard would
+    // otherwise fall back to default permissions, ignoring the profile.
+    // See #1142.
+    fetchProfiles().then((p) => {
+      dispatch({ type: "SET_PROFILES", profiles: p });
+      // Prefer an explicit prefill profile; otherwise use the server's active
+      // profile (`is_default: true`). If neither resolves, pass undefined so
+      // `fetchSettings` loads the unresolved global config.
+      const effectiveProfile =
+        prefill?.profile || p.find((x) => x.is_default)?.name || "";
+      fetchSettings(effectiveProfile || undefined).then((s) => {
+        if (!s) return;
         const sandbox = s.sandbox as Record<string, unknown> | undefined;
+        const session = s.session as Record<string, unknown> | undefined;
         const img = (sandbox?.default_image as string) || "";
         if (img) dispatch({ type: "SET_FIELD", field: "sandboxImage", value: img });
-      }
+        const env = Array.isArray(sandbox?.environment)
+          ? (sandbox?.environment as unknown[]).filter(
+              (v): v is string => typeof v === "string",
+            )
+          : [];
+        // Honor explicit prefill values so a caller that sets yoloMode/
+        // sandboxEnabled/tool isn't silently overridden by profile defaults.
+        // Mirrors the per-field guards `AgentStep.handleProfileChange` skips
+        // by going through the user-driven onChange path.
+        dispatch({
+          type: "APPLY_PROFILE_DEFAULTS",
+          yoloMode:
+            prefill?.yoloMode ??
+            ((session?.yolo_mode_default as boolean) ?? false),
+          sandboxEnabled:
+            prefill?.sandboxEnabled ??
+            ((sandbox?.enabled_by_default as boolean) ?? false),
+          tool: prefill?.tool || (session?.default_tool as string) || "",
+          extraEnv: env,
+          skipIfDirty: true,
+        });
+      });
     });
+    // prefill is captured at first render; we don't want to re-seed defaults
+    // (and stomp on user edits) if the parent re-renders with a new object
+    // identity.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const handleChange = useCallback((field: string, value: unknown) => {
